pyproject.toml
[project]
name = "pc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.0",
    "keyboard>=0.13.5",
    "pyserial>=3.5",
    "ftd2xx>=1.3.8",
    "numpy>=2.3.2",
    "pyqt5>=5.15.11",
    "pyqtgraph>=0.13.7",
]
